refactor(agents): move events into a seperate file
